axiom-driver: first implementation of axiom-bind API

+++ include/axiom_nic_types.h
@@ -23,7 +23,7 @@
 typedef uint8_t	            axiom_port_t;       /* Small message port */
 typedef uint8_t	            axiom_flag_t;       /* Small message flag */
 typedef uint8_t	            axiom_port_flag_t;  /* Small message port/flag */
-typedef uint16_t            axiom_small_len_t;  /* Small queue length */
+typedef uint32_t            axiom_small_len_t;  /* Small queue length */
 typedef uint8_t             axiom_node_id_t;    /* Node identifier */
 typedef uint8_t             axiom_msg_id_t;     /* Message identifier */
 typedef uint8_t	            axiom_if_id_t;      /* Interface identifier */
